[refactor] #62 : FSD 구조 레이어 추가
