+++ src/features/mapView/ui/index.ts
@@ -1,5 +1,4 @@
 export * from "./KakaoMapView";
-export * from "./MapHeader";
 export * from "./AddMemberBottomSheet";
 export * from "./MapDetailBottomSheet";
 export * from "./SnapBottomSheet";
