+++ src/shared/ui/index.ts
@@ -1,7 +1,4 @@
 export * from "./Button";
-export * from "./PlainHeader";
-export * from "./BackHeader";
-export * from "./CloseHeader";
 export * from "./LocationCard";
 export * from "./InputField";
 export * from "./PlaceCard";
