+++ src/widgets/Header/index.ts
@@ -1,0 +1,5 @@
+export * from "./ui/MapHeader";
+export * from "./ui/PlainHeader";
+export * from "./ui/BackHeader";
+export * from "./ui/CloseHeader";
+export * from "./model/types";
