+++ src/widgets/Header/model/types.ts
@@ -1,0 +1,18 @@
+export interface HeaderProps {
+  profileImg?: string;
+}
+
+export interface PlainHeaderProps {
+  title: string;
+  onBack?: () => void;
+  url?: string;
+}
+
+export interface BackHeaderProps {
+  url?: string;
+  onClick?: () => void;
+}
+
+export interface CloseHeaderProps {
+  url: string;
+}
